+++ sdk/communication/azure-communication-chat/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release History
 
+## 1.1.0-beta.3 (2021-09-15)
+- Version update.
+
 ## 1.1.0-beta.2 (2021-08-10)
 - Version update.
 - Fix `azure-communication-common` dependency issue.